deployment/requirements.txt
scikit_learn == 1.2.2
pandas
matplotlib
streamlit
Pillow
tensorflow == 2.15.0
numpy
matplotlib